RegExport 1.0
RegExport is a utility for Windows that allows users to quickly export selected Registry contents to a .reg file. The tool offers a number of hotkeys and a pop-up menu to make the process more efficient, allowing users to "live export" keys and entries to a .reg file opened in a Notepad window. The utility offers a number of features, including the ability to launch Regedit, select or create a .reg file, and tile Regedit and the .reg file's Notepad windows side by side on the screen. RegExport is free to download and use.